# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Order API Documentation",
        "description": "API documentation for your Food Ordering Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/user/orders": {
            "get": {
                "tags": [
                    "Orders (User)"
                ],
                "summary": "Get all orders",
                "operationId": "ae40bd7233855e22be39aa3f075919fd",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "Filter by orders id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders (User)"
                ],
                "summary": "Create new orders",
                "operationId": "e51a637b4fac0ebc9d5c6469d4af91c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "ordertype_id",
                                    "paymenttype_id",
                                    "stage_id",
                                    "subtotal",
                                    "discount",
                                    "delivery_fee",
                                    "service_fee",
                                    "total",
                                    "is_paid"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ordertype_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paymenttype_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "delivery_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "service_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "TX123456789"
                                    },
                                    "is_paid": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "order_note": {
                                        "type": "string",
                                        "example": "Please deliver between 5-6 PM"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-15T17:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/user/orders/{id}": {
            "get": {
                "tags": [
                    "Orders (User)"
                ],
                "summary": "Get order by ID",
                "operationId": "94db45f5c1f6e6ab146c2698e9a86dea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/user/payment-types": {
            "get": {
                "tags": [
                    "PaymentTypes (User)"
                ],
                "summary": "Get all payment types",
                "operationId": "8ec094db3f112f7da99a0a8b0569b370",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PaymentTypes list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}